Add files via upload
DefendHER is an Android Women Safety App that aims at empowering  women by providing them with a comprehensive safety solution. As we navigate the digital landscape of safety, DefendHER remains committed to its mission of ensuring the well-being and confidence of every user. It’s key features include Real-Time Tracking, Immediate Assistance through Emergency SOS Button, and Flexibility of SMS Alerts etc. It’s a Digital Guardian and a rapid response system that contributes to overall effectiveness and provides a highlighted security layer.